test-backend: Move nativeToCypher, cypherToNative and ResultObserver to their own files
The change was done to improve the readability of the main.js.

This pr also implement the cypherToNative for Float and Null and nativeToCypher for Bool and Int (which were not working properly)
