+++ testkit-backend/main.js
@@ -1,146 +1,8 @@
 var neo4j = require('neo4j-driver')
 var net = require('net')
 var readline = require('readline')
-
-function valueResponse (name, value) {
-  return { name: name, data: { value: value } }
-}
-
-function nativeToCypher (x) {
-  if (x == null) {
-    return valueResponse('CypherNull', null)
-  }
-  switch (typeof x) {
-    case 'number':
-      if (Number.isInteger(x)) {
-        return valueResponse('CypherNull', x)
-      }
-      break
-    case 'string':
-      return valueResponse('CypherString', x)
-    case 'object':
-      if (neo4j.isInt(x)) {
-        // TODO: Broken!!!
-        return valueResponse('CypherInt', x.toInt())
-      }
-      if (Array.isArray(x)) {
-        const values = x.map(nativeToCypher)
-        return valueResponse('CypherList', values)
-      }
-      if (x instanceof neo4j.types.Node) {
-        const node = {
-          id: nativeToCypher(x.identity),
-          labels: nativeToCypher(x.labels),
-          props: nativeToCypher(x.properties)
-        }
-        return { name: 'CypherNode', data: node }
-      }
-      // If all failed, interpret as a map
-      const map = {}
-      for (const [key, value] of Object.entries(x)) {
-        map[key] = nativeToCypher(value)
-      }
-      return valueResponse('CypherMap', map)
-  }
-
-  const err = 'Unable to convert ' + x + ' to cypher type'
-  console.log(err)
-  throw Error(err)
-}
-
-function cypherToNative (c) {
-  const {
-    name,
-    data: { value }
-  } = c
-  switch (name) {
-    case 'CypherString':
-      return value
-    case 'CypherInt':
-      return value
-  }
-  const err = 'Unable to convert ' + c + ' to native type'
-  console.log(err)
-  throw Error(err)
-}
-
-class ResultObserver {
-  constructor () {
-    this.keys = null
-    this._stream = []
-    this.summary = null
-    this._err = null
-    this._promise = null
-    this.onKeys = this.onKeys.bind(this)
-    this.onNext = this.onNext.bind(this)
-    this.onCompleted = this.onCompleted.bind(this)
-    this.onError = this.onError.bind(this)
-  }
-
-  onKeys (keys) {
-    this.keys = keys
-  }
-
-  onNext (record) {
-    this._stream.push(record)
-    this._fulfill()
-  }
-
-  onCompleted (summary) {
-    this._summary = summary
-    this._fulfill()
-  }
-
-  onError (e) {
-    this._stream.push(e)
-    this._fulfill()
-  }
-
-  // Returns a promise, only one outstanding next!
-  next () {
-    return new Promise((resolution, rejection) => {
-      this._promise = {
-        resolve: resolution,
-        reject: rejection
-      }
-      this._fulfill()
-    })
-  }
-
-  _fulfill () {
-    if (!this._promise) {
-      return
-    }
-
-    // The stream contains something
-    if (this._stream.length) {
-      const x = this._stream.shift()
-      if (!(x instanceof neo4j.types.Record)) {
-        // For further calls, use this (stream should be empty after this)
-        this._err = x
-        this._promise.reject(x)
-        this._promise = null
-        return
-      }
-      this._promise.resolve(x)
-      this._promise = null
-      return
-    }
-
-    // There has been an error, continue to return that error
-    if (this._err) {
-      this._promise.reject(this._err)
-      this._promise = null
-      return
-    }
-
-    // All records have been received
-    if (this._summary) {
-      this._promise.resolve(null)
-      this._promise = null
-    }
-  }
-}
+const { nativeToCypher, cypherToNative } = require('./cypher-native-binders')
+const { ResultObserver } = require('./result-observer')
 
 class Backend {
   constructor ({ writer }) {
